Move some assignments back into Pod::Html
In the subroutine exported by Pod::Html::Auxiliary, little more should
be done than to wrap around GetOptions and provide a usage message if
needed.  This subroutine should not need to be aware of the globals
already under assignment in Pod::Html.

Hence, rename that subroutine to process_command_line(), limit what it
does, and have it return a reference to a hash.

In Pod::Html::pod2html(), introduce process_options() to handle the
initial merging of the data in $opts into $globals.

Signed-off-by: James E Keenan <jkeenan@cpan.org>